+++ src/components/Card/Card.tsx
@@ -5,8 +5,8 @@ import {Action, DisableableAction, contentContextTypes} from '../../types';
 import {buttonFrom} from '../Button';
 import ButtonGroup from '../ButtonGroup';
 
-import Header from './Header';
-import Section from './Section';
+import Header from './components/Header';
+import Section from './components/Section';
 import * as styles from './Card.scss';
 
 export interface Props {