+++ src/components/ColorPicker/components/AlphaPicker/index.ts
@@ -1,0 +1,4 @@
+import AlphaPicker from './AlphaPicker';
+
+export {Props} from './AlphaPicker';
+export default AlphaPicker;