+++ src/components/ColorPicker/components/Slidable/index.ts
@@ -1,0 +1,4 @@
+import Slidable from './Slidable';
+
+export {Props, Position} from './Slidable';
+export default Slidable;