+++ src/components/ColorPicker/index.ts
@@ -1,7 +1,7 @@
 import ColorPicker from './ColorPicker';
 
 export * from './types';
-export * from './color-utilities';
+export * from './utilities/color-utilities';
 
 export {Props} from './ColorPicker';
 export default ColorPicker;