+++ src/components/DataTable/components/Cell/index.ts
@@ -1,0 +1,4 @@
+import Cell from './Cell';
+
+export {Props} from './Cell';
+export default Cell;