+++ src/components/FormLayout/components/Group/index.ts
@@ -1,0 +1,4 @@
+import Group from './Group';
+
+export {Props} from './Group';
+export default Group;