+++ src/components/Page/components/Action/index.ts
@@ -1,0 +1,3 @@
+import Action from './Action';
+
+export default Action;