+++ src/components/Page/tests/Header.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {mountWithAppProvider} from '../../../../tests/utilities';
 import {LinkAction} from '../../../types';
-import Header from '../../Page/Header';
+import Header from '../components/Header';
 import Breadcrumbs from '../../Breadcrumbs';
 
 const breadcrumbs: LinkAction[] = [