+++ src/components/Tabs/components/List/index.ts
@@ -1,0 +1,4 @@
+import List from './List';
+
+export {Props} from './List';
+export default List;