+++ src/components/Tabs/components/Panel/index.ts
@@ -1,0 +1,4 @@
+import Panel from './Panel';
+
+export {Props} from './Panel';
+export default Panel;